requirements.txt
opencv-python>=4.7.0
Pillow>=9.4.0
numpy>=1.23.0
sounddevice
scipy 